rename gluu-pep to gluu-uma-pep, #258
