+++ kong/plugins/gluu-uma-pep/access.lua
@@ -1,13 +1,13 @@
 local constants = require "kong.constants"
 local oxd = require "oxdweb"
-local helper = require "kong.plugins.gluu-pep.helper"
+local helper = require "kong.plugins.gluu-uma-pep.helper"
 
 -- we don't store our token in lrucache - we don't want it be pushed out
 local access_token
 local access_token_expire = 0
 local EXPIRE_DELTA = 10
 
-local PLUGINNAME = "gluu-pep"
+local PLUGINNAME = "gluu-uma-pep"
 
 local lrucache = require "resty.lrucache"
 -- it can be shared by all the requests served by each nginx worker process:
