+++ t/specs/gluu-uma-pep/test_spec.lua
@@ -7,7 +7,7 @@ local JSON = require"JSON"
 
 local host_git_root = os.getenv"HOST_GIT_ROOT"
 local git_root = os.getenv"GIT_ROOT"
-local test_root = host_git_root .. "/t/specs/gluu-pep"
+local test_root = host_git_root .. "/t/specs/gluu-uma-pep"
 
 local function setup(model)
     _G.ctx = {}
@@ -37,7 +37,7 @@ local function setup(model)
     kong_utils.docker_unique_network()
     kong_utils.kong_postgress_custom_plugins{
         plugins = {
-            ["gluu-pep"] = host_git_root .. "/kong/plugins/gluu-pep",
+            ["gluu-uma-pep"] = host_git_root .. "/kong/plugins/gluu-uma-pep",
         },
         modules = {
             ["oxdweb.lua"] = host_git_root .. "/third-party/oxd-web-lua/oxdweb.lua",
@@ -105,7 +105,7 @@ local function configure_plugin(create_service_response, plugin_config)
     plugin_config.oxd_id = register_site_response.oxd_id
 
     local payload = {
-        name = "gluu-pep",
+        name = "gluu-uma-pep",
         config = plugin_config,
         service_id = create_service_response.id,
     }
@@ -123,7 +123,7 @@ local function configure_plugin(create_service_response, plugin_config)
 end
 
 
-test("gluu-pep", function()
+test("gluu-uma-pep", function()
     setup("oxd-model1.lua")
 
     local create_service_response = configure_service_route()
@@ -197,7 +197,7 @@ test("gluu-pep", function()
     ctx.print_logs = false
 end)
 
-test("gluu-pep allow_unprotected_path = true", function()
+test("gluu-uma-pep allow_unprotected_path = true", function()
     setup("oxd-model2.lua")
 
     local create_service_response = configure_service_route()
